requirements.txt
fastapi~=0.128.1
uvicorn
requests~=2.31.0
glom~=25.12.0
python-dotenv~=1.0.0
load_dotenv
httpx~=0.28.1
google-cloud-storage
google-cloud-secret-manager
google-cloud-bigquery
google-cloud-pubsub
google-auth
pydantic~=2.12.5
six~=1.16.0
google-cloud-iam-credentials